Crate exact_cover

An exact cover problem is a mathematical problem that consists of a universe of “items” and a set of “options” which are sets of items. A solution to the problem is a set of options whose union is the universe of items and whose pairwise intersections are all empty. In other words, a solution must cover each item exactly once.

Any problem that involves placing a finite set of objects in a finite set of possible positions under some non-overlapping constraint can likely be reduced to (or formulated in terms of) an exact cover problem. One example is the N Queens puzzle, which, for example, tries to place 8 queens on a standard 8x8 chessboard such that no two queens can block each other’s movement. Another example is Sudoku, in which each cell in a 9x9 grid must be filled with one of the numbers 1-9 such that no column, row, or 3x3 sub-grid contains two of the same number.

Donald Knuth’s Algorithm X can find all solutions of any exact cover problem. From a high level, Algorithm X is simply a backtracking exhaustive search algorithm. At a lower level, Algorithm X cleverly manipulates linked lists to make the backtracking more efficient. This module implements Algorithm X but separates the two levels: the backtracking search is implemented by the ExactCoverProblem trait, while the lower level linked list data structure and operations are implemented by the Dlx struct. This split arrangement gives the implementor more flexibility to customize the search algorithm for the specific problem at hand. A ready-to-use implementation of ExactCoverProblem built on Dlx can be found at MrvExactCoverSearch, and its code can be helpful as a starting point for custom implementations.

The algorithms and data structures in this module, as well as the terminology of “options” and “items”, come from Donald Knuth’s The Art of Computer Programming, section 7.2.2, https://www-cs-faculty.stanford.edu/~knuth/fasc5c.ps.gz.

Quick Start Roadmap

Map your problem to an exact cover problem.

Build the exact cover formulation with Dlx::new and Dlx::add_option.

Pass the dlx instance to MrvExactCoverSearch::new.

Call MrvExactCoverSearch::search.

Check if a solution was found with MrvExactCoverSearch::current_solution.

The solution is simply a list of options and will need to be mapped back into your problem.

Optional: if the last search returned true, steps 4-6 can be repeated to find more solutions.

Structs

Dlx

“Dancing links” data structure described by Knuth for solving exact cover problems. See module documentation for additional details.

DlxOption

Opaque handle used by Dlx to refer to an option.

MrvExactCoverSearch

An exact cover search implementation using the minimum-remaining-values (MRV) heuristic, i.e. items with the fewest available options are selected first.

Traits

ExactCoverProblem

A generalized exact cover problem. See module documentation for additional details.