Struct exact_cover::Dlx

pub struct Dlx { /* private fields */ }

“Dancing links” data structure described by Knuth for solving exact cover problems. See module documentation for additional details.

This implementation supports optional secondary items which can be covered 0 or 1 times in a solution, as the search algorithm needs no modifications to support this extension.

Some of the methods in this type have usage requirements that, if violated, are likely to silently corrupt the internal data structures leading to incorrect results and possibly panics. These methods include sanity checks some of which are implemented with debug_assert! for performance reasons. Thus, it is recommended to test with debug_assertions enabled.

Implementations

impl Dlx

pub fn new(n_primary: usize, n_secondary: usize) -> Self

Creates a Dlx instance with the specified number of primary (required) and secondary (optional) items, but no options. Primary items are automatically numbered starting at 1, and secondary item numbers follow consecutively after primary items.

pub fn add_option<'a, I: IntoIterator<Item = &'a usize>>(&mut self, option: I)

Adds option represented as a sorted (ascending order) collection of unique item indices.

Item indices start at 1 (index 0 is reserved). Empty options are silently ignored. Panics if the above requirements are not met. Do not call this method when a search for a solution is in progress.

pub fn select_item(&mut self, item: usize)

Covers item and sets it as the current item.

Covering an item removes it from the set of outstanding uncovered items and removes all options that contain it from the set of available options.

The selected item must not already be covered, and there must not be a current item already set.

Note that this method does not actually select an option, so in general, the next step after calling this method should be to select_option or undo_item if no more options are available.

pub fn undo_item(&mut self)

Undoes select_item.

pub fn next_option(&self, prev: Option<DlxOption>) -> Option<DlxOption>

Returns the next available option that includes the current item. Returns the first available option if prev is None. Returns None if prev is the last available option for the item.

This method must not be called if there is no current item.

pub fn select_option(&mut self, option: DlxOption)

Adds option to the candidate solution and covers all uncovered items included in the option.

This method must only be called when there is a current item set, and the option argument must include the current item. Violating these conditions will likely corrupt the data structure. When called it unsets the current item.

pub fn try_undo_option(&mut self) -> Option<DlxOption>

Undoes the most recent select_option and returns the option that was deselected, or does nothing and returns None if there was nothing to undo.

This method must not be called when there is a current item. It restores the previous current item if an option was undone.

pub fn primary_items(&self) -> impl Iterator<Item = usize> + '_

Returns an iterator over the remaining uncovered primary items.

pub fn options_len(&self, item: usize) -> usize

Returns the number of currently available options that include item.

pub fn current_solution(&self) -> Option<&[DlxOption]>

Returns a slice of options in the currently selected solution, or None if not in a solved state. Use option_items to get the items in the option.

pub fn option_items(&self, option: DlxOption) -> &[usize]

Returns a slice of item indices in the given option.